splits utils into h and cpp; moves some parts into distance_matrix.cpp
